feat(test-audit): split tests/test_memory_server.py monolith into domain-specific modules (Closes #857)
Split the 2035-line test monolith into three domain-specific files and
added contract-style mock helpers to conftest.py:

- test_memory_server_pure.py — pure function tests (RRF merge, temporal
  scoring, formatting, cosine sim, pgvector parsing, excluded-tags filter,
  recall dedup)
- test_memory_server_store.py — store/classifier/auto-links/embedding tests
- test_memory_server_recall.py — recall pipeline tests (link expansion,
  lifecycle filters, outcome memory ID)

conftest.py now includes:
- _FakeTable / _FakeExecute / _FakeQueryResult contract-style test doubles
  for reduced mock call-chain assertions
- make_mock_client factory fixture

All 105 tests pass. No test coverage lost — each class moved in full.
