--- tests/test_memory_server.py
@@ -1,2035 +1,0 @@
-"""Unit tests for mcp-memory/server.py — pure functions + mocked async.
-
-Covers Memory 2.0 core: temporal scoring, RRF merge, formatting, auto-linking.
-"""
-
-from __future__ import annotations
-
-import json
-import os
-import sys
-import types
-from datetime import datetime, timezone, timedelta
-from pathlib import Path
-from unittest.mock import MagicMock
-
-import pytest
-
-# ---------------------------------------------------------------------------
-# Mock external dependencies before importing server.py
-# server.py imports mcp, supabase, httpx at module level — we stub them out
-# so tests run without installing the full MCP SDK.
-# ---------------------------------------------------------------------------
-
-# Create stub modules for mcp.*
-_mcp_types = types.ModuleType("mcp.types")
-_mcp_types.CallToolResult = MagicMock
-
-
-class _FakeTextContent:
-    """Minimal TextContent replica so tests can read back .text on returns.
-
-    The real mcp.types.TextContent is a pydantic model; tests don't need
-    validation, just attribute access for the error-path checks below.
-    """
-
-    def __init__(self, type: str = "text", text: str = ""):
-        self.type = type
-        self.text = text
-
-
-_mcp_types.TextContent = _FakeTextContent
-_mcp_types.Tool = MagicMock
-
-
-def _noop_decorator(*args, **kwargs):
-    """Return a no-op decorator that passes the function through."""
-
-    def decorator(fn):
-        return fn
-
-    return decorator
-
-
-class _FakeServer:
-    """Minimal stub that supports @server.list_tools() and @server.call_tool() decorators."""
-
-    def __init__(self, *args, **kwargs):
-        pass
-
-    def list_tools(self):
-        return _noop_decorator()
-
-    def call_tool(self):
-        return _noop_decorator()
-
-
-_mcp_server = types.ModuleType("mcp.server")
-_mcp_server.Server = _FakeServer
-
-_mcp_server_stdio = types.ModuleType("mcp.server.stdio")
-_mcp_server_stdio.stdio_server = MagicMock
-
-_mcp = types.ModuleType("mcp")
-
-for mod_name, mod in [
-    ("mcp", _mcp),
-    ("mcp.types", _mcp_types),
-    ("mcp.server", _mcp_server),
-    ("mcp.server.stdio", _mcp_server_stdio),
-]:
-    sys.modules.setdefault(mod_name, mod)
-
-# Stub supabase only if it's not actually installed. Blindly stubbing with
-# setdefault would shadow a real install from other test modules that run
-# in the same session (e.g. test_agents_smoke.py imports real `Client`).
-try:
-    import supabase  # noqa: F401
-except ImportError:
-    sys.modules["supabase"] = types.ModuleType("supabase")
-
-# Stub httpx only if it's not actually installed. A blind setdefault
-# would shadow a real install for the rest of the pytest session (e.g.
-# test_agents_smoke.py needs real `httpx.get` to monkey-patch the
-# GitHub client).
-try:
-    import httpx  # noqa: F401
-except ImportError:
-    sys.modules["httpx"] = types.ModuleType("httpx")
-
-# Stub dotenv
-_dotenv = types.ModuleType("dotenv")
-_dotenv.load_dotenv = lambda *a, **kw: None
-sys.modules.setdefault("dotenv", _dotenv)
-
-# Add mcp-memory to path
-sys.path.insert(0, str(Path(__file__).resolve().parent.parent / "mcp-memory"))
-
-# Env vars (server reads them for client init)
-os.environ.setdefault("SUPABASE_URL", "https://test.supabase.co")
-os.environ.setdefault("SUPABASE_KEY", "test-key")
-
-from server import (
-    _rrf_merge,
-    _apply_temporal_scoring,
-    _format_memories,
-    _create_auto_links,
-    _expand_with_links,
-    _handle_store,
-    TEMPORAL_HALF_LIVES,
-    SUPERSEDE_SIM_THRESHOLD,
-    MAX_AUTO_LINKS,
-)
-import server as server_module
-
-
-# ---------------------------------------------------------------------------
-# _rrf_merge
-# ---------------------------------------------------------------------------
-
-
-class TestRRFMerge:
-    """Reciprocal Rank Fusion merges two ranked lists."""
-
-    def test_empty_inputs(self):
-        result = _rrf_merge([], [], limit=5)
-        assert result == []
-
-    def test_single_list(self):
-        rows = [{"id": "a", "name": "mem1"}, {"id": "b", "name": "mem2"}]
-        result = _rrf_merge(rows, [], limit=5)
-        assert len(result) == 2
-        assert result[0]["id"] == "a"  # rank 0 scores higher
-        assert result[1]["id"] == "b"
-
-    def test_overlap_boosts_score(self):
-        """Items in both lists should score higher than items in one."""
-        sem = [{"id": "a", "name": "shared"}, {"id": "b", "name": "sem-only"}]
-        kw = [{"id": "a", "name": "shared"}, {"id": "c", "name": "kw-only"}]
-        result = _rrf_merge(sem, kw, limit=5)
-
-        # 'a' appears in both → highest score; use _final_score (always present,
-        # unlike _rrf_score which is dual-hit only in the canonical rrf_merge).
-        assert result[0]["id"] == "a"
-        assert result[0]["_final_score"] > result[1]["_final_score"]
-
-    def test_limit_respected(self):
-        rows = [{"id": str(i), "name": f"m{i}"} for i in range(10)]
-        result = _rrf_merge(rows, [], limit=3)
-        assert len(result) == 3
-
-    def test_score_calculation(self):
-        """Verify RRF score formula: 1/(k + rank)."""
-        k = 60
-        rows = [{"id": "a", "name": "x"}]
-        result = _rrf_merge(rows, [], limit=5, k=k)
-        expected = 1.0 / (k + 0)
-        # Single-hit rows use _final_score (unified key); _rrf_score is dual-hit only.
-        assert abs(result[0]["_final_score"] - expected) < 1e-10
-
-    def test_score_both_lists(self):
-        """Score = sum of 1/(k+rank) from each list."""
-        k = 60
-        sem = [{"id": "a", "name": "x"}]
-        kw = [{"id": "z", "name": "y"}, {"id": "a", "name": "x"}]  # 'a' at rank 1 in kw
-        result = _rrf_merge(sem, kw, limit=5, k=k)
-
-        a_score = next(r for r in result if r["id"] == "a")
-        expected = 1.0 / (k + 0) + 1.0 / (k + 1)  # rank 0 in sem + rank 1 in kw
-        assert abs(a_score["_rrf_score"] - expected) < 1e-10
-
-    def test_name_fallback_for_id(self):
-        """Rows without 'id' use 'name' as identifier."""
-        rows = [{"name": "alpha"}, {"name": "beta"}]
-        result = _rrf_merge(rows, [], limit=5)
-        assert len(result) == 2
-
-
-# ---------------------------------------------------------------------------
-# _apply_temporal_scoring
-# ---------------------------------------------------------------------------
-
-
-class TestTemporalScoring:
-    """Temporal scoring re-ranks by recency × access boost."""
-
-    @staticmethod
-    def _make_row(mem_type="decision", days_ago=0, accessed_days_ago=None, rrf=0.5):
-        now = datetime.now(timezone.utc)
-        updated = (now - timedelta(days=days_ago)).isoformat()
-        accessed = (
-            (now - timedelta(days=accessed_days_ago)).isoformat()
-            if accessed_days_ago is not None
-            else None
-        )
-        return {
-            "type": mem_type,
-            "updated_at": updated,
-            "last_accessed_at": accessed,
-            "_rrf_score": rrf,
-        }
-
-    def test_recent_ranks_higher(self):
-        old = self._make_row(days_ago=30)
-        new = self._make_row(days_ago=1)
-        result = _apply_temporal_scoring([old, new])
-        assert result[0]["_temporal_score"] > result[1]["_temporal_score"]
-
-    def test_same_age_equal_scores(self):
-        a = self._make_row(days_ago=5, rrf=0.5)
-        b = self._make_row(days_ago=5, rrf=0.5)
-        result = _apply_temporal_scoring([a, b])
-        assert abs(result[0]["_temporal_score"] - result[1]["_temporal_score"]) < 1e-6
-
-    def test_access_boost(self):
-        """Recently accessed memory gets boosted."""
-        accessed = self._make_row(days_ago=10, accessed_days_ago=1)
-        not_accessed = self._make_row(days_ago=10)
-        result = _apply_temporal_scoring([not_accessed, accessed])
-        assert result[0]["_temporal_score"] > result[1]["_temporal_score"]
-
-    def test_type_half_lives(self):
-        """Project memories decay faster than feedback memories."""
-        project = self._make_row(mem_type="project", days_ago=14)  # half-life 7d
-        feedback = self._make_row(mem_type="feedback", days_ago=14)  # half-life 90d
-        result = _apply_temporal_scoring([project, feedback])
-        # feedback should score higher — 14 days is well within its 90d half-life
-        assert result[0]["type"] == "feedback"
-
-    def test_rrf_weight_preserved(self):
-        """Higher RRF score leads to higher temporal score, all else equal."""
-        high_rrf = self._make_row(rrf=0.9, days_ago=5)
-        low_rrf = self._make_row(rrf=0.1, days_ago=5)
-        result = _apply_temporal_scoring([low_rrf, high_rrf])
-        assert result[0]["_temporal_score"] > result[1]["_temporal_score"]
-
-    def test_handles_missing_timestamps(self):
-        """Gracefully handles rows with missing/invalid timestamps."""
-        row = {"type": "decision", "updated_at": "", "last_accessed_at": None, "_rrf_score": 0.5}
-        result = _apply_temporal_scoring([row])
-        assert "_temporal_score" in result[0]
-        assert result[0]["_temporal_score"] > 0
-
-    def test_zero_days_ago(self):
-        """Just-updated memory gets near-full score."""
-        row = self._make_row(days_ago=0, rrf=1.0)
-        result = _apply_temporal_scoring([row])
-        # recency ≈ 1.0, so temporal_score ≈ rrf * 1.0 * access_factor
-        assert result[0]["_temporal_score"] >= 0.9
-
-    def test_sorted_descending(self):
-        rows = [
-            self._make_row(days_ago=30, rrf=0.3),
-            self._make_row(days_ago=1, rrf=0.8),
-            self._make_row(days_ago=10, rrf=0.5),
-        ]
-        result = _apply_temporal_scoring(rows)
-        scores = [r["_temporal_score"] for r in result]
-        assert scores == sorted(scores, reverse=True)
-
-    # -- #240: confidence entrenchment multiplier ---------------------------
-
-    def test_confidence_high_outranks_low(self):
-        """Higher-confidence row ranks above lower-confidence at equal age/RRF."""
-        high = self._make_row(days_ago=5, rrf=0.5)
-        high["id"] = "high"
-        high["confidence"] = 1.0
-        low = self._make_row(days_ago=5, rrf=0.5)
-        low["id"] = "low"
-        low["confidence"] = 0.5
-        result = _apply_temporal_scoring([low, high])
-        assert result[0]["id"] == "high"
-        # score(conf=1.0) / score(conf=0.5) = 1.0 / 0.75 exactly.
-        ratio = result[0]["_temporal_score"] / result[1]["_temporal_score"]
-        assert abs(ratio - (1.0 / 0.75)) < 1e-9
-
-    def test_confidence_null_treated_as_1(self):
-        """Legacy rows (no confidence) score same as confidence=1.0 (no regression)."""
-        null_row = self._make_row(days_ago=5, rrf=0.5)
-        null_row["id"] = "null"
-        full_row = self._make_row(days_ago=5, rrf=0.5)
-        full_row["id"] = "full"
-        full_row["confidence"] = 1.0
-        result = _apply_temporal_scoring([null_row, full_row])
-        s_null = next(r["_temporal_score"] for r in result if r["id"] == "null")
-        s_full = next(r["_temporal_score"] for r in result if r["id"] == "full")
-        assert abs(s_null - s_full) < 1e-9
-
-    def test_confidence_zero_gets_floor(self):
-        """confidence=0 → score = CONFIDENCE_FLOOR * score(confidence=1.0) (soft floor)."""
-        from server import CONFIDENCE_FLOOR
-
-        zero = self._make_row(days_ago=5, rrf=0.5)
-        zero["id"] = "zero"
-        zero["confidence"] = 0.0
-        full = self._make_row(days_ago=5, rrf=0.5)
-        full["id"] = "full"
-        full["confidence"] = 1.0
-        result = _apply_temporal_scoring([zero, full])
-        s_zero = next(r["_temporal_score"] for r in result if r["id"] == "zero")
-        s_full = next(r["_temporal_score"] for r in result if r["id"] == "full")
-        assert abs(s_zero / s_full - CONFIDENCE_FLOOR) < 1e-9
-
-
-# ---------------------------------------------------------------------------
-# _format_memories
-# ---------------------------------------------------------------------------
-
-
-class TestFormatMemories:
-    """Memory formatting for LLM output."""
-
-    def test_basic_format(self):
-        mem = {
-            "name": "test_mem",
-            "type": "decision",
-            "project": "jarvis",
-            "description": "A test memory",
-            "content": "Some content here",
-            "tags": ["tag1", "tag2"],
-            "updated_at": "2026-04-09T12:00:00+00:00",
-        }
-        result = _format_memories([mem])
-        assert len(result) == 1
-        assert "## test_mem (decision, jarvis)" in result[0]
-        assert "[tag1, tag2]" in result[0]
-        assert "*A test memory*" in result[0]
-        assert "Some content here" in result[0]
-
-    def test_global_project(self):
-        mem = {
-            "name": "x",
-            "type": "user",
-            "project": None,
-            "description": "",
-            "content": "c",
-            "tags": [],
-        }
-        result = _format_memories([mem])
-        assert "(user, global)" in result[0]
-
-    def test_no_tags(self):
-        mem = {"name": "x", "type": "user", "project": None, "description": "", "content": "c"}
-        result = _format_memories([mem])
-        # Header should NOT contain tag brackets when tags are missing
-        header_line = result[0].split("\n")[0]
-        assert "] " not in header_line or header_line.endswith(")")
-
-    def test_empty_tags_list(self):
-        mem = {
-            "name": "x",
-            "type": "user",
-            "project": None,
-            "description": "",
-            "content": "c",
-            "tags": [],
-        }
-        result = _format_memories([mem])
-        header_line = result[0].split("\n")[0]
-        # Empty tags list should not produce brackets
-        assert "[]" not in header_line
-
-    def test_link_info(self):
-        mem = {
-            "name": "linked_mem",
-            "type": "decision",
-            "project": "jarvis",
-            "description": "desc",
-            "content": "body",
-            "tags": [],
-            "link_type": "related",
-            "link_strength": 0.75,
-            "updated_at": "2026-04-09",
-        }
-        result = _format_memories([mem], link_info=True)
-        assert "← related (0.75)" in result[0]
-
-    def test_link_info_not_shown_without_flag(self):
-        mem = {
-            "name": "x",
-            "type": "decision",
-            "project": None,
-            "description": "",
-            "content": "c",
-            "tags": [],
-            "link_type": "related",
-            "link_strength": 0.75,
-        }
-        result = _format_memories([mem], link_info=False)
-        assert "← related" not in result[0]
-
-    def test_multiple_memories(self):
-        mems = [
-            {
-                "name": f"m{i}",
-                "type": "project",
-                "project": "j",
-                "description": "",
-                "content": f"c{i}",
-                "tags": [],
-            }
-            for i in range(5)
-        ]
-        result = _format_memories(mems)
-        assert len(result) == 5
-
-    # Brief mode — Phase 7.2. One-line rows so bulk/auto-injection sites
-    # don't pay full-content budget.
-
-    def test_brief_basic_layout(self):
-        mem = {
-            "name": "foo",
-            "type": "feedback",
-            "project": "jarvis",
-            "tags": ["a", "b"],
-            "description": "hello world",
-            "similarity": 0.42,
-            "content": "MUST_NOT_APPEAR",
-        }
-        result = _format_memories([mem], brief=True)
-        assert result == ["- foo [feedback/jarvis] [a, b] (sim 0.42): hello world — id=?"]
-        assert "MUST_NOT_APPEAR" not in result[0]
-
-    def test_brief_global_scope(self):
-        mem = {"name": "g", "type": "user", "project": None, "description": "d"}
-        result = _format_memories([mem], brief=True)
-        assert result[0] == "- g [user/global]: d — id=?"
-
-    def test_brief_temporal_score_leads_when_present(self):
-        # `_temporal_score` is what drives actual ordering after
-        # _apply_temporal_scoring. It must appear first so the shown score
-        # matches the displayed rank; the retrieval signal (rrf/sim) trails
-        # as provenance.
-        mem = {
-            "name": "t",
-            "type": "decision",
-            "project": "jarvis",
-            "description": "x",
-            "_temporal_score": 0.0456,
-            "_rrf_score": 0.0333,
-        }
-        result = _format_memories([mem], brief=True)
-        assert "(score 0.046; rrf 0.033)" in result[0]
-
-    def test_brief_rrf_wins_over_similarity(self):
-        mem = {
-            "name": "r",
-            "type": "decision",
-            "description": "x",
-            "_rrf_score": 0.05,
-            "similarity": 0.9,
-        }
-        result = _format_memories([mem], brief=True)
-        assert "rrf 0.050" in result[0]
-        assert "sim" not in result[0]
-
-    def test_brief_link_info(self):
-        mem = {
-            "name": "l",
-            "type": "decision",
-            "project": "jarvis",
-            "description": "d",
-            "similarity": 0.5,
-            "link_type": "related",
-            "link_strength": 0.75,
-        }
-        result = _format_memories([mem], link_info=True, brief=True)
-        assert "← related (0.75)" in result[0]
-        assert result[0].startswith("- l [decision/jarvis]")
-
-    def test_brief_no_score_fields(self):
-        mem = {"name": "n", "type": "reference", "project": None, "description": "d"}
-        result = _format_memories([mem], brief=True)
-        assert result[0] == "- n [reference/global]: d — id=?"
-
-    def test_brief_empty_description(self):
-        # Migration-target memories carry empty descriptions — brief still
-        # surfaces the name (no crash, trailing `: ` is intentional).
-        mem = {"name": "bare", "type": "decision", "project": "jarvis"}
-        result = _format_memories([mem], brief=True)
-        assert result[0] == "- bare [decision/jarvis]:  — id=?"
-
-
-# ---------------------------------------------------------------------------
-# _create_auto_links (async, mocked Supabase)
-# ---------------------------------------------------------------------------
-
-
-class TestCreateAutoLinks:
-    """Auto-linking creates memory_links entries based on similarity.
-
-    Phase 2b changed the contract: links are always created as 'related'
-    first, then the classifier (or legacy fallback) decides whether to
-    upgrade to 'supersedes' / mark expired. We assert against the FIRST
-    upsert call (the related batch) where it matters.
-    """
-
-    @pytest.fixture
-    def mock_client(self):
-        client = MagicMock()
-        client.table.return_value.upsert.return_value.execute.return_value = MagicMock(data=[])
-        # Hydration query returns no rows (test path skips classifier anyway).
-        client.table.return_value.select.return_value.in_.return_value.execute.return_value = (
-            MagicMock(data=[])
-        )
-        return client
-
-    def _first_links_upsert(self, mock_client):
-        """Return the first list-of-links arg passed to upsert."""
-        for call in mock_client.table.return_value.upsert.call_args_list:
-            arg = call[0][0]
-            if isinstance(arg, list):
-                return arg
-        return []
-
-    @pytest.mark.asyncio
-    async def test_creates_related_links(self, mock_client):
-        similar = [
-            {"id": "target-1", "type": "project", "similarity": 0.70},
-            {"id": "target-2", "type": "project", "similarity": 0.65},
-        ]
-        await _create_auto_links(mock_client, "source-id", similar, mem_type="project")
-
-        links = self._first_links_upsert(mock_client)
-        assert len(links) == 2
-        assert all(l["link_type"] == "related" for l in links)
-        assert links[0]["strength"] == 0.70
-        assert links[1]["strength"] == 0.65
-
-    @pytest.mark.asyncio
-    async def test_legacy_fallback_supersedes_same_type(self, mock_client, monkeypatch):
-        """When the classifier is unavailable (no API key, no candidate),
-        the legacy heuristic fires: same type + sim >= 0.85 → supersede."""
-        monkeypatch.delenv("ANTHROPIC_API_KEY", raising=False)
-        similar = [
-            {
-                "id": "old-decision",
-                "type": "decision",
-                "similarity": SUPERSEDE_SIM_THRESHOLD + 0.05,
-            },
-        ]
-        await _create_auto_links(mock_client, "new-decision", similar, mem_type="decision")
-
-        # The legacy fallback updates memories.superseded_by on the target.
-        update_calls = [
-            c
-            for c in mock_client.table.return_value.update.call_args_list
-            if c[0][0].get("superseded_by") == "new-decision"
-        ]
-        assert len(update_calls) == 1
-
-    @pytest.mark.asyncio
-    async def test_no_supersession_when_below_threshold(self, mock_client, monkeypatch):
-        """Even same-type, below SUPERSEDE_SIM_THRESHOLD → no supersession."""
-        monkeypatch.delenv("ANTHROPIC_API_KEY", raising=False)
-        similar = [
-            {"id": "target", "type": "project", "similarity": 0.70},
-        ]
-        await _create_auto_links(mock_client, "source", similar, mem_type="project")
-
-        update_calls = [
-            c
-            for c in mock_client.table.return_value.update.call_args_list
-            if c[0][0].get("superseded_by") == "source"
-        ]
-        assert update_calls == []
-
-    @pytest.mark.asyncio
-    async def test_max_links_limit(self, mock_client):
-        similar = [{"id": f"t-{i}", "type": "project", "similarity": 0.70} for i in range(10)]
-        await _create_auto_links(mock_client, "source", similar, mem_type="project")
-
-        links = self._first_links_upsert(mock_client)
-        assert len(links) == MAX_AUTO_LINKS
-
-    @pytest.mark.asyncio
-    async def test_empty_similar_rows(self, mock_client):
-        await _create_auto_links(mock_client, "source", [], mem_type="project")
-        # No links to insert → the memory_links upsert should not fire.
-        link_calls = [c for c in mock_client.table.call_args_list if c[0][0] == "memory_links"]
-        assert link_calls == []
-
-    @pytest.mark.asyncio
-    async def test_swallows_exceptions(self, mock_client):
-        """Fire-and-forget: errors don't propagate."""
-        mock_client.table.side_effect = Exception("DB error")
-        # Should not raise
-        await _create_auto_links(
-            mock_client, "source", [{"id": "t", "type": "p", "similarity": 0.7}], "project"
-        )
-
-
-# ---------------------------------------------------------------------------
-# Phase 2b classifier — _apply_classifier_decision routing + queue writes
-# ---------------------------------------------------------------------------
-
-from server import _apply_classifier_decision, CLASSIFIER_APPLY_THRESHOLD  # noqa: E402
-from classifier import ClassifierDecision  # noqa: E402
-
-
-class TestApplyClassifierDecision:
-    """Routing of classifier decisions: which DB mutations fire when."""
-
-    @pytest.fixture
-    def mock_client(self):
-        client = MagicMock()
-        # Per-table mocks so we can filter calls by table name.
-        # Each call to client.table("foo") returns the same mock instance
-        # for "foo", but a *different* mock for "bar" — that's how we tell
-        # an insert into memory_review_queue from one into memory_links.
-        client._tables = {}
-
-        def _get_table(name):
-            if name not in client._tables:
-                t = MagicMock()
-                # update().eq().is_().execute() returns truthy .data so the
-                # rowcount check in _apply_classifier_decision sees a hit.
-                t.update.return_value.eq.return_value.is_.return_value.execute.return_value = (
-                    MagicMock(data=[{"id": "row"}])
-                )
-                t.insert.return_value.execute.return_value = MagicMock()
-                t.upsert.return_value.execute.return_value = MagicMock()
-                client._tables[name] = t
-            return client._tables[name]
-
-        client.table.side_effect = _get_table
-        return client
-
-    def _update_calls(self, mock_client, key: str, table: str = "memories"):
-        """Find update() calls on a specific table whose payload contains a key."""
-        t = mock_client._tables.get(table)
-        if t is None:
-            return []
-        return [c for c in t.update.call_args_list if isinstance(c[0][0], dict) and key in c[0][0]]
-
-    def _queue_inserts(self, mock_client):
-        """Find insert calls into memory_review_queue specifically."""
-        t = mock_client._tables.get("memory_review_queue")
-        if t is None:
-            return []
-        return t.insert.call_args_list
-
-    def _link_upserts(self, mock_client, link_type: str | None = None):
-        """Find upsert calls into memory_links, optionally filtered by link_type."""
-        t = mock_client._tables.get("memory_links")
-        if t is None:
-            return []
-        calls = t.upsert.call_args_list
-        if link_type is None:
-            return calls
-        out = []
-        for c in calls:
-            payload = c[0][0]
-            if isinstance(payload, dict) and payload.get("link_type") == link_type:
-                out.append(c)
-        return out
-
-    @pytest.mark.asyncio
-    async def test_high_confidence_update_marks_superseded(self, mock_client):
-        decision = ClassifierDecision(
-            decision="UPDATE",
-            target_id="old-id",
-            confidence=0.95,
-            reasoning="refines target",
-        )
-        neighbors = [{"id": "old-id", "name": "old", "similarity": 0.82}]
-        await _apply_classifier_decision(mock_client, "new-id", decision, neighbors)
-
-        # The target was marked superseded.
-        sup_calls = self._update_calls(mock_client, "superseded_by")
-        assert len(sup_calls) == 1
-        assert sup_calls[0][0][0]["superseded_by"] == "new-id"
-
-        # The related-link was upgraded to a `supersedes` link in memory_links.
-        sup_links = self._link_upserts(mock_client, link_type="supersedes")
-        assert len(sup_links) == 1
-        link_payload = sup_links[0][0][0]
-        assert link_payload["source_id"] == "new-id"
-        assert link_payload["target_id"] == "old-id"
-
-        # And the decision was recorded with status=auto_applied.
-        inserts = self._queue_inserts(mock_client)
-        assert len(inserts) == 1
-        payload = inserts[0][0][0]
-        assert payload["decision"] == "UPDATE"
-        assert payload["status"] == "auto_applied"
-        assert payload["target_id"] == "old-id"
-
-    @pytest.mark.asyncio
-    async def test_high_confidence_delete_sets_expired(self, mock_client):
-        decision = ClassifierDecision(
-            decision="DELETE",
-            target_id="old-id",
-            confidence=0.92,
-            reasoning="negates target",
-        )
-        neighbors = [{"id": "old-id", "name": "old", "similarity": 0.85}]
-        await _apply_classifier_decision(mock_client, "new-id", decision, neighbors)
-
-        exp_calls = self._update_calls(mock_client, "expired_at")
-        assert len(exp_calls) == 1
-
-        inserts = self._queue_inserts(mock_client)
-        assert inserts[0][0][0]["decision"] == "DELETE"
-        assert inserts[0][0][0]["status"] == "auto_applied"
-
-    @pytest.mark.asyncio
-    async def test_low_confidence_update_queues_pending(self, mock_client):
-        """Low confidence: do NOT mutate target, write queue entry as pending."""
-        decision = ClassifierDecision(
-            decision="UPDATE",
-            target_id="old-id",
-            confidence=CLASSIFIER_APPLY_THRESHOLD - 0.1,
-            reasoning="ambiguous",
-        )
-        neighbors = [{"id": "old-id", "name": "old", "similarity": 0.78}]
-        await _apply_classifier_decision(mock_client, "new-id", decision, neighbors)
-
-        sup_calls = self._update_calls(mock_client, "superseded_by")
-        assert sup_calls == []  # nothing mutated
-
-        inserts = self._queue_inserts(mock_client)
-        payload = inserts[0][0][0]
-        assert payload["status"] == "pending"
-        assert payload["applied_at"] is None
-
-    @pytest.mark.asyncio
-    async def test_noop_records_decision_no_mutation(self, mock_client):
-        decision = ClassifierDecision(
-            decision="NOOP",
-            target_id=None,
-            confidence=0.9,
-            reasoning="redundant",
-        )
-        neighbors = [{"id": "x", "name": "x", "similarity": 0.9}]
-        await _apply_classifier_decision(mock_client, "new-id", decision, neighbors)
-
-        # No supersede / expired mutations.
-        assert self._update_calls(mock_client, "superseded_by") == []
-        assert self._update_calls(mock_client, "expired_at") == []
-        # Decision still recorded for audit.
-        inserts = self._queue_inserts(mock_client)
-        assert len(inserts) == 1
-        assert inserts[0][0][0]["decision"] == "NOOP"
-
-    @pytest.mark.asyncio
-    async def test_high_confidence_add_no_queue_entry(self, mock_client):
-        """ADD with high confidence is the trivial case — don't pollute queue."""
-        decision = ClassifierDecision(
-            decision="ADD",
-            target_id=None,
-            confidence=0.95,
-            reasoning="genuinely new",
-        )
-        neighbors = [{"id": "x", "name": "x", "similarity": 0.76}]
-        await _apply_classifier_decision(mock_client, "new-id", decision, neighbors)
-
-        assert self._queue_inserts(mock_client) == []
-
-    @pytest.mark.asyncio
-    async def test_hallucinated_target_id_refused(self, mock_client):
-        """Model returned an id we never showed it → refuse to mutate."""
-        decision = ClassifierDecision(
-            decision="UPDATE",
-            target_id="never-existed",
-            confidence=0.95,
-            reasoning="...",
-        )
-        neighbors = [{"id": "real-id", "name": "real", "similarity": 0.85}]
-        await _apply_classifier_decision(mock_client, "new-id", decision, neighbors)
-
-        assert self._update_calls(mock_client, "superseded_by") == []
-        # Still recorded, but with status=pending and target_id=None.
-        inserts = self._queue_inserts(mock_client)
-        payload = inserts[0][0][0]
-        assert payload["status"] == "pending"
-        assert payload["target_id"] is None
-
-
-# ---------------------------------------------------------------------------
-# _expand_with_links (async, mocked Supabase)
-# ---------------------------------------------------------------------------
-
-
-class TestExpandWithLinks:
-    """Graph traversal fetches 1-hop linked memories."""
-
-    @pytest.fixture
-    def mock_client(self):
-        client = MagicMock()
-        return client
-
-    @pytest.mark.asyncio
-    async def test_returns_linked_memories(self, mock_client):
-        mock_client.rpc.return_value.execute.return_value = MagicMock(
-            data=[
-                {
-                    "id": "linked-1",
-                    "name": "linked_mem",
-                    "link_type": "related",
-                    "link_strength": 0.75,
-                },
-            ]
-        )
-        result = await _expand_with_links(mock_client, ["source-id"])
-        assert len(result) == 1
-        assert result[0]["id"] == "linked-1"
-
-    @pytest.mark.asyncio
-    async def test_empty_when_no_links(self, mock_client):
-        mock_client.rpc.return_value.execute.return_value = MagicMock(data=[])
-        result = await _expand_with_links(mock_client, ["source-id"])
-        assert result == []
-
-    @pytest.mark.asyncio
-    async def test_swallows_exceptions(self, mock_client):
-        mock_client.rpc.side_effect = Exception("DB error")
-        result = await _expand_with_links(mock_client, ["source-id"])
-        assert result == []
-
-    @pytest.mark.asyncio
-    async def test_passes_memory_ids_to_rpc(self, mock_client):
-        mock_client.rpc.return_value.execute.return_value = MagicMock(data=[])
-        await _expand_with_links(mock_client, ["id-1", "id-2"])
-        # include_unreviewed=False is the always-gate default (#552); the
-        # parameter is now plumbed through so future call sites can opt in.
-        mock_client.rpc.assert_called_once_with(
-            "get_linked_memories",
-            {
-                "memory_ids": ["id-1", "id-2"],
-                "link_types": None,
-                "show_history": False,
-                "include_unreviewed": False,
-            },
-        )
-
-    @pytest.mark.asyncio
-    async def test_passes_show_history_to_rpc(self, mock_client):
-        mock_client.rpc.return_value.execute.return_value = MagicMock(data=[])
-        await _expand_with_links(mock_client, ["id-1"], show_history=True)
-        mock_client.rpc.assert_called_once_with(
-            "get_linked_memories",
-            {
-                "memory_ids": ["id-1"],
-                "link_types": None,
-                "show_history": True,
-                "include_unreviewed": False,
-            },
-        )
-
-    @pytest.mark.asyncio
-    async def test_passes_include_unreviewed_to_rpc(self, mock_client):
-        """When opt-in flag is set, it must reach the RPC body."""
-        mock_client.rpc.return_value.execute.return_value = MagicMock(data=[])
-        await _expand_with_links(mock_client, ["id-1"], include_unreviewed=True)
-        mock_client.rpc.assert_called_once_with(
-            "get_linked_memories",
-            {
-                "memory_ids": ["id-1"],
-                "link_types": None,
-                "show_history": False,
-                "include_unreviewed": True,
-            },
-        )
-
-
-# ---------------------------------------------------------------------------
-# Recall pipeline dedup regression (bidirectional link bug)
-# ---------------------------------------------------------------------------
-
-
-class TestRecallLinkedDedup:
-    """Regression: bidirectional links should not produce duplicate linked memories."""
-
-    def test_dedup_within_linked_results(self):
-        """If _expand_with_links returns the same memory twice (bidirectional links),
-        the recall pipeline should deduplicate to keep only one."""
-        # Simulate: main results have IDs ["main-1", "main-2"]
-        # Linked results return "linked-A" twice (from both link directions)
-        main_ids = {"main-1", "main-2"}
-        linked = [
-            {"id": "linked-A", "name": "mem_a", "link_type": "related", "link_strength": 0.71},
-            {"id": "linked-A", "name": "mem_a", "link_type": "related", "link_strength": 0.63},
-            {"id": "linked-B", "name": "mem_b", "link_type": "related", "link_strength": 0.65},
-        ]
-
-        # Reproduce the dedup logic from _hybrid_recall
-        found_ids = set(main_ids)
-        seen_linked: set[str] = set()
-        unique_linked = []
-        for r in linked:
-            rid = r.get("id")
-            if rid not in found_ids and rid not in seen_linked:
-                seen_linked.add(rid)
-                unique_linked.append(r)
-
-        assert len(unique_linked) == 2
-        assert unique_linked[0]["id"] == "linked-A"
-        assert unique_linked[0]["link_strength"] == 0.71  # first (strongest) kept
-        assert unique_linked[1]["id"] == "linked-B"
-
-    def test_main_results_excluded_from_linked(self):
-        """Linked results that duplicate main results should be excluded."""
-        main_ids = {"id-1"}
-        linked = [
-            {"id": "id-1", "name": "same_as_main", "link_type": "related", "link_strength": 0.80},
-            {"id": "id-2", "name": "different", "link_type": "related", "link_strength": 0.70},
-        ]
-
-        found_ids = set(main_ids)
-        seen_linked: set[str] = set()
-        unique_linked = []
-        for r in linked:
-            rid = r.get("id")
-            if rid not in found_ids and rid not in seen_linked:
-                seen_linked.add(rid)
-                unique_linked.append(r)
-
-        assert len(unique_linked) == 1
-        assert unique_linked[0]["id"] == "id-2"
-
-
-# ---------------------------------------------------------------------------
-# Phase 2c: memory_store must reject writes missing source_provenance
-# ---------------------------------------------------------------------------
-
-
-class TestHandleStoreProvenance:
-    """Phase 2c — every memory write carries a namespaced source_provenance.
-
-    The MCP boundary rejects missing/blank values with a readable error so
-    callers don't hit a raw NOT NULL violation from Postgres downstream.
-    """
-
-    @pytest.fixture(autouse=True)
-    def _patch_client(self, monkeypatch):
-        """Swap _get_client for a mock so validation failures don't need live DB."""
-        self.client = MagicMock()
-        monkeypatch.setattr(server_module, "_get_client", lambda: self.client)
-
-    @pytest.mark.asyncio
-    async def test_rejects_missing_provenance(self):
-        result = await _handle_store(
-            {
-                "type": "project",
-                "name": "test_missing",
-                "content": "test content",
-                # no source_provenance
-            }
-        )
-        assert len(result) == 1
-        assert "source_provenance is required" in result[0].text
-        # Validation fired before any DB access.
-        self.client.table.assert_not_called()
-
-    @pytest.mark.asyncio
-    async def test_rejects_blank_provenance(self):
-        result = await _handle_store(
-            {
-                "type": "project",
-                "name": "test_blank",
-                "content": "test content",
-                "source_provenance": "   ",
-            }
-        )
-        assert "source_provenance is required" in result[0].text
-        self.client.table.assert_not_called()
-
-    @pytest.mark.asyncio
-    async def test_rejects_none_provenance(self):
-        result = await _handle_store(
-            {
-                "type": "project",
-                "name": "test_none",
-                "content": "test content",
-                "source_provenance": None,
-            }
-        )
-        assert "source_provenance is required" in result[0].text
-        self.client.table.assert_not_called()
-
-    @pytest.mark.asyncio
-    async def test_provenance_stripped_before_persist(self, monkeypatch):
-        """Accepted provenance is trimmed — no leading/trailing whitespace
-        leaks into the DB row, keeping audit queries clean."""
-
-        # Short-circuit embedding so we don't need Voyage env/network.
-        # Signature accepts **kwargs so #242's model= param doesn't break it.
-        async def _fake_embed(_text, **_kwargs):
-            return None
-
-        monkeypatch.setattr(server_module, "_embed", _fake_embed)
-
-        # project="jarvis" takes the upsert branch. Rig the chain to return
-        # a stored id so _handle_store completes without errors.
-        tbl = MagicMock()
-        tbl.upsert.return_value.execute.return_value = MagicMock(data=[{"id": "stored-1"}])
-        self.client.table.return_value = tbl
-
-        await _handle_store(
-            {
-                "type": "project",
-                "name": "test_strip",
-                "content": "test content",
-                "project": "jarvis",
-                "source_provenance": "  skill:test  ",
-            }
-        )
-
-        upsert_calls = tbl.upsert.call_args_list
-        assert upsert_calls, "expected at least one upsert call"
-        data_arg = upsert_calls[-1][0][0]
-        assert data_arg["source_provenance"] == "skill:test"
-
-
-class TestHandleStoreStructuredResponse:
-    """#658: success-path returns JSON, not prose.
-
-    Iter:13 of an AFK chain reported a fabricated "dup-detector ≥0.75 sim
-    block" against a same-name upsert and gave up after 4 retries. The
-    diagnosis (issue #658 comment) showed the store call never landed —
-    some other failure mode surfaced as prose the agent rationalised into
-    a plausible-sounding semantic-similarity rejection. The fix is making
-    success unambiguous: callers parse a JSON envelope with a `stored`
-    boolean instead of inferring from a sentence.
-
-    Invariants these tests pin:
-    - `(project, name)` is the only identity. Same-name writes are
-      idempotent — no similarity threshold ever blocks.
-    - `consolidation_candidates` and `classifier_pending` are advisory
-      fields — they describe the post-store landscape, never gate it.
-    - Consolidation phrasing is neutral (no `⚠`, no "hint" framing) so
-      the LLM doesn't mistake it for a rejection signal.
-    """
-
-    @pytest.fixture(autouse=True)
-    def _patch_client(self, monkeypatch):
-        self.client = MagicMock()
-        monkeypatch.setattr(server_module, "_get_client", lambda: self.client)
-
-        # Default: no embedding (skip sim-search path). Tests that need
-        # consolidation override this with a non-None vector.
-        async def _no_embed(_text):
-            return {}
-
-        monkeypatch.setattr(server_module, "_compute_write_embeddings", _no_embed)
-
-        # Neutralise fire-and-forget classifier dispatch so the test event
-        # loop doesn't surface unawaited-coroutine warnings from real calls.
-        async def _noop_links(*_a, **_k):
-            return None
-
-        monkeypatch.setattr(server_module, "_create_auto_links", _noop_links)
-
-    @pytest.mark.asyncio
-    async def test_project_scoped_upsert_returns_structured_json(self):
-        tbl = MagicMock()
-        tbl.upsert.return_value.execute.return_value = MagicMock(data=[{"id": "mem-uuid-1"}])
-        self.client.table.return_value = tbl
-
-        result = await _handle_store(
-            {
-                "type": "project",
-                "name": "test_struct_project",
-                "content": "test content",
-                "project": "jarvis",
-                "source_provenance": "session:test",
-            }
-        )
-
-        assert len(result) == 1
-        body = json.loads(result[0].text)
-        assert body["stored"] is True
-        assert body["action"] == "saved"
-        assert body["memory_id"] == "mem-uuid-1"
-        assert body["project"] == "jarvis"
-        # No embedding ⇒ no sim search ran ⇒ both advisory fields are empty.
-        assert body["consolidation_candidates"] == []
-        assert body["classifier_pending"] is False
-        # Human-readable summary still present for LLM readers.
-        assert "test_struct_project" in body["message"]
-        assert "saved" in body["message"]
-
-    @pytest.mark.asyncio
-    async def test_global_new_returns_action_created(self):
-        tbl = MagicMock()
-        # SELECT branch: row does not exist yet.
-        select_chain = tbl.select.return_value.eq.return_value.is_.return_value
-        select_chain.limit.return_value.execute.return_value = MagicMock(data=[])
-        # INSERT returns the new id.
-        tbl.insert.return_value.execute.return_value = MagicMock(data=[{"id": "mem-new"}])
-        self.client.table.return_value = tbl
-
-        result = await _handle_store(
-            {
-                "type": "feedback",
-                "name": "test_struct_global_new",
-                "content": "x",
-                "source_provenance": "session:test",
-            }
-        )
-
-        body = json.loads(result[0].text)
-        assert body["stored"] is True
-        assert body["action"] == "created"
-        assert body["memory_id"] == "mem-new"
-        assert body["project"] == "global"
-
-    @pytest.mark.asyncio
-    async def test_global_existing_returns_action_updated(self):
-        tbl = MagicMock()
-        select_chain = tbl.select.return_value.eq.return_value.is_.return_value
-        select_chain.limit.return_value.execute.return_value = MagicMock(
-            data=[{"id": "mem-existing"}]
-        )
-        tbl.update.return_value.eq.return_value.execute.return_value = MagicMock(data=[])
-        self.client.table.return_value = tbl
-
-        result = await _handle_store(
-            {
-                "type": "feedback",
-                "name": "test_struct_global_update",
-                "content": "x",
-                "source_provenance": "session:test",
-            }
-        )
-
-        body = json.loads(result[0].text)
-        assert body["stored"] is True
-        assert body["action"] == "updated"
-        assert body["memory_id"] == "mem-existing"
-        assert body["project"] == "global"
-
-    @pytest.mark.asyncio
-    async def test_consolidation_uses_neutral_phrasing(self, monkeypatch):
-        """No `⚠` glyph and no 'hint' framing — the field is advisory.
-
-        iter:13's confabulation was prompted by `⚠ Consolidation hint:` —
-        the warning glyph reads as a rejection signal to a tired LLM.
-        Drop the glyph, drop the word 'hint'; prefix with `info:` so the
-        framing is purely descriptive.
-        """
-        # Real-ish embedding so the sim-search block runs.
-        async def _real_embed(_text):
-            return {"embedding": [0.1] * 512}
-
-        monkeypatch.setattr(server_module, "_compute_write_embeddings", _real_embed)
-
-        tbl = MagicMock()
-        tbl.upsert.return_value.execute.return_value = MagicMock(data=[{"id": "mem-uuid-2"}])
-        self.client.table.return_value = tbl
-
-        # RPC returns ≥CONSOLIDATION_COUNT candidates above
-        # CONSOLIDATION_SIM_THRESHOLD (=0.80) so the hint fires.
-        self.client.rpc.return_value.execute.return_value = MagicMock(
-            data=[
-                {"id": "sib-1", "name": "sibling_a", "similarity": 0.85},
-                {"id": "sib-2", "name": "sibling_b", "similarity": 0.82},
-                {"id": "sib-3", "name": "sibling_c", "similarity": 0.81},
-            ]
-        )
-
-        result = await _handle_store(
-            {
-                "type": "feedback",
-                "name": "test_struct_consolidation",
-                "content": "content",
-                "project": "jarvis",
-                "source_provenance": "session:test",
-            }
-        )
-
-        body = json.loads(result[0].text)
-        # Store succeeded — never blocked.
-        assert body["stored"] is True
-        assert body["action"] == "saved"
-        # Neutral phrasing — no warning glyph, no "hint" framing.
-        assert "⚠" not in body["message"]
-        assert "hint" not in body["message"].lower()
-        assert "info:" in body["message"].lower()
-        # Structured field carries the candidate names for programmatic
-        # consumers; prose message is for human/LLM readers only.
-        assert body["consolidation_candidates"] == [
-            "sibling_a",
-            "sibling_b",
-            "sibling_c",
-        ]
-        # similar_rows was non-empty ⇒ classifier task dispatched.
-        assert body["classifier_pending"] is True
-
-
-# ---------------------------------------------------------------------------
-# #242: dual-embedding machinery — column/RPC mapping + dual-write
-# ---------------------------------------------------------------------------
-
-from server import (  # noqa: E402
-    _model_slot,
-    _embed_upsert_fields,
-    _compute_write_embeddings,
-)
-
-
-class TestModelSlotMapping:
-    """#242: the model → column/RPC table drives both read and write paths.
-
-    If this mapping drifts, writes land in the wrong column and reads
-    query the wrong RPC — a silent data-integrity bug. Lock it down.
-    """
-
-    def test_voyage_3_lite_maps_to_v1_column(self):
-        slot = _model_slot("voyage-3-lite")
-        assert slot["embedding_column"] == "embedding"
-        assert slot["rpc"] == "match_memories"
-
-    def test_voyage_3_maps_to_v2_column(self):
-        slot = _model_slot("voyage-3")
-        assert slot["embedding_column"] == "embedding_v2"
-        assert slot["rpc"] == "match_memories_v2"
-
-    def test_unknown_model_falls_back_to_legacy(self):
-        """Misconfigured EMBEDDING_MODEL_PRIMARY must degrade to legacy,
-        never raise at runtime — writes continue, bug is visible in metadata."""
-        slot = _model_slot("nonexistent-model")
-        assert slot["embedding_column"] == "embedding"
-        assert slot["rpc"] == "match_memories"
-
-    def test_upsert_fields_shape(self):
-        fields = _embed_upsert_fields([0.1, 0.2], "voyage-3-lite")
-        assert fields == {
-            "embedding": [0.1, 0.2],
-            "embedding_model": "voyage-3-lite",
-            "embedding_version": "v2",
-        }
-
-    def test_upsert_fields_v2(self):
-        fields = _embed_upsert_fields([0.3, 0.4], "voyage-3")
-        assert fields == {
-            "embedding_v2": [0.3, 0.4],
-            "embedding_model_v2": "voyage-3",
-            "embedding_version_v2": "v2",
-        }
-
-    def test_upsert_fields_unknown_returns_empty(self):
-        """Unknown model → no fields (no silent corruption of other columns)."""
-        assert _embed_upsert_fields([0.1], "no-such-model") == {}
-
-
-class TestDualEmbedWrite:
-    """#242: when SECONDARY is set, writes compute both embeddings and
-    populate both columns. Unset → identical to pre-#242 (single-write)."""
-
-    @pytest.mark.asyncio
-    async def test_secondary_unset_single_write(self, monkeypatch):
-        """Zero-change default: SECONDARY unset → only PRIMARY columns written."""
-        calls: list[dict] = []
-
-        async def fake_embed(text, input_type="document", model=None):
-            calls.append({"model": model})
-            return [0.1, 0.2, 0.3]
-
-        monkeypatch.setattr(server_module, "_embed", fake_embed)
-        monkeypatch.setattr(server_module, "EMBEDDING_MODEL_PRIMARY", "voyage-3-lite")
-        monkeypatch.setattr(server_module, "EMBEDDING_MODEL_SECONDARY", None)
-
-        fields = await _compute_write_embeddings("canonical text")
-
-        assert "embedding" in fields
-        assert "embedding_model" in fields
-        assert fields["embedding_model"] == "voyage-3-lite"
-        assert "embedding_v2" not in fields
-        assert len(calls) == 1
-        assert calls[0]["model"] == "voyage-3-lite"
-
-    @pytest.mark.asyncio
-    async def test_secondary_set_dual_write(self, monkeypatch):
-        """SECONDARY=voyage-3 → both columns populated, two embed calls."""
-        calls: list[dict] = []
-
-        async def fake_embed(text, input_type="document", model=None):
-            calls.append({"model": model})
-            # Return dim-shaped distinct vectors so we can assert routing.
-            return [0.1] * 512 if model == "voyage-3-lite" else [0.9] * 1024
-
-        monkeypatch.setattr(server_module, "_embed", fake_embed)
-        monkeypatch.setattr(server_module, "EMBEDDING_MODEL_PRIMARY", "voyage-3-lite")
-        monkeypatch.setattr(server_module, "EMBEDDING_MODEL_SECONDARY", "voyage-3")
-
-        fields = await _compute_write_embeddings("canonical text")
-
-        assert fields["embedding"] == [0.1] * 512
-        assert fields["embedding_v2"] == [0.9] * 1024
-        assert fields["embedding_model"] == "voyage-3-lite"
-        assert fields["embedding_model_v2"] == "voyage-3"
-        assert {c["model"] for c in calls} == {"voyage-3-lite", "voyage-3"}
-
-    @pytest.mark.asyncio
-    async def test_secondary_failure_single_leg(self, monkeypatch):
-        """PRIMARY succeeds but SECONDARY embed returns None → write PRIMARY
-        only. Missing v2 is recoverable via backfill; corrupt row is not."""
-
-        async def fake_embed(text, input_type="document", model=None):
-            if model == "voyage-3":
-                return None
-            return [0.1] * 512
-
-        monkeypatch.setattr(server_module, "_embed", fake_embed)
-        monkeypatch.setattr(server_module, "EMBEDDING_MODEL_PRIMARY", "voyage-3-lite")
-        monkeypatch.setattr(server_module, "EMBEDDING_MODEL_SECONDARY", "voyage-3")
-
-        fields = await _compute_write_embeddings("canonical text")
-        assert fields["embedding"] == [0.1] * 512
-        assert "embedding_v2" not in fields
-
-    @pytest.mark.asyncio
-    async def test_primary_failure_no_write(self, monkeypatch):
-        """PRIMARY embed fails → no embed fields at all. Row still saves
-        with text content, _backfill_missing_embeddings picks it up later."""
-
-        async def fake_embed(text, input_type="document", model=None):
-            return None
-
-        monkeypatch.setattr(server_module, "_embed", fake_embed)
-        monkeypatch.setattr(server_module, "EMBEDDING_MODEL_PRIMARY", "voyage-3-lite")
-        monkeypatch.setattr(server_module, "EMBEDDING_MODEL_SECONDARY", "voyage-3")
-
-        fields = await _compute_write_embeddings("canonical text")
-        assert fields == {}
-
-    @pytest.mark.asyncio
-    async def test_secondary_equals_primary_no_duplicate_call(self, monkeypatch):
-        """If SECONDARY accidentally equals PRIMARY, don't waste an API call."""
-        calls: list[dict] = []
-
-        async def fake_embed(text, input_type="document", model=None):
-            calls.append({"model": model})
-            return [0.1] * 512
-
-        monkeypatch.setattr(server_module, "_embed", fake_embed)
-        monkeypatch.setattr(server_module, "EMBEDDING_MODEL_PRIMARY", "voyage-3-lite")
-        monkeypatch.setattr(server_module, "EMBEDDING_MODEL_SECONDARY", "voyage-3-lite")
-
-        fields = await _compute_write_embeddings("canonical text")
-        assert len(calls) == 1
-        assert "embedding" in fields
-        # Only PRIMARY columns; no redundant duplicate write.
-        assert "embedding_v2" not in fields
-
-
-# =========================================================================
-# Known unknowns — retrieval gaps + unsatisfied queries (#249)
-# =========================================================================
-
-from server import (
-    _cosine_sim,
-    _parse_pgvector,
-    _upsert_known_unknown,
-    _resolve_known_unknowns,
-)
-
-
-class TestCosineSim:
-    """Unit tests for cosine similarity function."""
-
-    def test_empty_vectors(self):
-        assert _cosine_sim([], []) == 0.0
-
-    def test_none_vectors(self):
-        assert _cosine_sim(None, [1.0, 0.0]) == 0.0
-        assert _cosine_sim([1.0, 0.0], None) == 0.0
-        assert _cosine_sim(None, None) == 0.0
-
-    def test_identical_unit_vectors(self):
-        v = [1.0, 0.0]
-        assert _cosine_sim(v, v) == pytest.approx(1.0)
-
-    def test_orthogonal_vectors(self):
-        assert _cosine_sim([1.0, 0.0], [0.0, 1.0]) == pytest.approx(0.0)
-
-    def test_opposite_vectors(self):
-        assert _cosine_sim([1.0, 0.0], [-1.0, 0.0]) == pytest.approx(-1.0)
-
-    def test_scaled_same_direction(self):
-        # Cosine sim is invariant to magnitude
-        assert _cosine_sim([1.0, 1.0], [2.0, 2.0]) == pytest.approx(1.0)
-
-    def test_zero_magnitude_vector(self):
-        assert _cosine_sim([0.0, 0.0], [1.0, 1.0]) == 0.0
-
-    def test_length_mismatch_returns_zero(self):
-        # Dim mismatch must not silently truncate via zip — returns 0.0.
-        assert _cosine_sim([1.0, 0.0, 0.0], [1.0, 0.0]) == 0.0
-        assert _cosine_sim([1.0] * 512, [1.0] * 1024) == 0.0
-
-
-class TestParsePgvector:
-    """supabase-py returns pgvector columns as JSON strings, not lists.
-
-    Regression guard for the known-unknowns dedup/resolution bug where raw
-    string embeddings hit the `_cosine_sim` length-mismatch guard and
-    silently scored 0.
-    """
-
-    def test_list_passes_through(self):
-        v = [0.1, 0.2, 0.3]
-        assert _parse_pgvector(v) is v
-
-    def test_none_returns_none(self):
-        assert _parse_pgvector(None) is None
-
-    def test_json_string_parses_to_list(self):
-        parsed = _parse_pgvector("[0.1, 0.2, 0.3]")
-        assert parsed == [0.1, 0.2, 0.3]
-
-    def test_malformed_string_returns_none(self):
-        assert _parse_pgvector("not-json") is None
-        assert _parse_pgvector("[0.1, 0.2,") is None
-
-    def test_non_list_json_returns_none(self):
-        # A valid JSON number or object is not a vector — reject.
-        assert _parse_pgvector("42") is None
-        assert _parse_pgvector('{"x": 1}') is None
-
-    def test_unsupported_type_returns_none(self):
-        assert _parse_pgvector(42) is None  # type: ignore[arg-type]
-
-    def test_string_embedding_yields_nonzero_similarity(self):
-        """Core regression: feed a JSON-string embedding through the parser,
-        then _cosine_sim against an identical list must score ~1.0. Before
-        the fix, len("[0.1,...]") != len([0.1,...]) → guard returned 0.0."""
-        vec = [0.1] * 512
-        stored_as_string = json.dumps(vec)  # what supabase-py actually returns
-        parsed = _parse_pgvector(stored_as_string)
-        assert parsed is not None
-        assert len(parsed) == 512
-        assert _cosine_sim(vec, parsed) == pytest.approx(1.0)
-
-
-class TestKnownUnknowns:
-    """Unit tests for known_unknowns insertion + dedup + resolution."""
-
-    @pytest.mark.asyncio
-    async def test_known_unknowns_insert_on_low_sim(self):
-        """When recall finds top_similarity < 0.45, log as known unknown.
-        No existing row with matching query → insert (not update)."""
-        mock_client = MagicMock()
-
-        # select().eq().eq().limit().execute() returns no data
-        # (fallback path — no embedding, because 3 dims ≠ 512)
-        mock_select_chain = MagicMock()
-        mock_select_chain.eq.return_value.eq.return_value.limit.return_value.execute.return_value = MagicMock(
-            data=[]
-        )
-
-        mock_insert = MagicMock()
-        mock_update = MagicMock()
-
-        mock_table = MagicMock()
-        mock_table.select.return_value = mock_select_chain
-        mock_table.insert.return_value = mock_insert
-        mock_table.update.return_value = mock_update
-        mock_client.table.return_value = mock_table
-
-        # 3-dim embedding gets coerced to None by dim guard → fallback path
-        await _upsert_known_unknown(
-            mock_client,
-            query="what is the meaning of life",
-            query_embedding=[0.1, 0.2, 0.3],
-            top_similarity=0.3,
-            top_memory_id="mem-123",
-            context={"project": "jarvis"},
-        )
-
-        # Assert insert was called with correct payload, and update was NOT
-        mock_insert.execute.assert_called_once()
-        insert_payload = mock_table.insert.call_args.args[0]
-        assert insert_payload["query"] == "what is the meaning of life"
-        assert insert_payload["top_similarity"] == 0.3
-        assert insert_payload["top_memory_id"] == "mem-123"
-        assert insert_payload["query_embedding"] is None  # dim-guarded
-        assert not mock_update.execute.called
-
-    @pytest.mark.asyncio
-    async def test_known_unknowns_dedup_increments_hit_count(self):
-        """If similar query exists (cosine > 0.9), increment hit_count instead of insert.
-        Validates that the select includes hit_count and the update uses the stored value."""
-        mock_client = MagicMock()
-
-        # Use 512-dim embeddings (matches schema vector(512)).
-        # Two nearly identical vectors → cosine ≈ 1.0 > 0.9 → dedup fires.
-        existing_embedding = [0.10] * 512
-        similar_embedding = [0.11] * 512
-
-        # Mock the select chain: table().select("id, query_embedding, hit_count").eq().execute()
-        mock_select_return = MagicMock()
-        mock_select_return.eq.return_value.execute.return_value = MagicMock(
-            data=[{"id": "uk-1", "query_embedding": existing_embedding, "hit_count": 5}]
-        )
-
-        mock_update_return = MagicMock()
-        mock_insert_return = MagicMock()
-
-        mock_table = MagicMock()
-        mock_table.select.return_value = mock_select_return
-        mock_table.update.return_value = mock_update_return
-        mock_table.insert.return_value = mock_insert_return
-        mock_client.table.return_value = mock_table
-
-        await _upsert_known_unknown(
-            mock_client,
-            query="what is the meaning of existence",
-            query_embedding=similar_embedding,
-            top_similarity=0.35,
-            top_memory_id="mem-456",
-        )
-
-        # Select must include hit_count so the increment reflects stored value
-        select_cols = mock_table.select.call_args.args[0]
-        assert "hit_count" in select_cols
-
-        # Update was called once with hit_count = 5 + 1 = 6 (not 1 + 1 = 2)
-        mock_table.update.assert_called_once()
-        update_payload = mock_table.update.call_args.args[0]
-        assert update_payload["hit_count"] == 6
-
-        # Insert was NOT called
-        assert not mock_insert_return.execute.called
-
-    @pytest.mark.asyncio
-    async def test_known_unknowns_resolution_on_store(self):
-        """When a memory is stored, resolve open unknowns with cosine > 0.7."""
-        mock_client = MagicMock()
-        # Existing open unknown
-        unknown_embedding = [0.5, 0.5, 0.0]
-
-        # Setup the mock chain for table().select().eq().execute()
-        mock_select = MagicMock()
-        mock_eq = MagicMock()
-        mock_eq.execute.return_value = MagicMock(
-            data=[{"id": "uk-2", "query_embedding": unknown_embedding}]
-        )
-        mock_select.eq.return_value = mock_eq
-
-        # For the update chain
-        mock_update = MagicMock()
-        mock_update_eq = MagicMock()
-        mock_update_eq.execute.return_value = MagicMock()
-        mock_update.eq.return_value = mock_update_eq
-
-        # Setup table() to return different objects for select vs update
-        def table_side_effect(table_name):
-            if table_name == "known_unknowns":
-                result = MagicMock()
-                result.select.return_value = mock_select
-                result.update.return_value = mock_update
-                return result
-            return MagicMock()
-
-        mock_client.table.side_effect = table_side_effect
-
-        # New memory with matching embedding
-        memory_embedding = [0.6, 0.55, 0.1]
-        await _resolve_known_unknowns(mock_client, memory_embedding, "mem-789")
-
-        # Verify update was called with status='resolved'
-        assert mock_update.eq.called
-
-
-# ---------------------------------------------------------------------------
-# #284: memory_recall must exclude soft-deleted + superseded + expired rows
-# ---------------------------------------------------------------------------
-
-
-class TestRecallLifecycleFilters:
-    """Regression (Osasuwu/jarvis#284): memory_recall surfaced soft-deleted
-    rows even when show_history=false, because the SQL RPCs and the Python
-    keyword-only fallback forgot parts of the lifecycle filter. Lock the
-    contract at both layers: the RPC params we send, and the query-builder
-    filters we apply.
-    """
-
-    @staticmethod
-    def _fluent_query():
-        """MagicMock that returns itself from every builder method — lets
-        us assert on .is_ / .or_ / .eq / .limit calls without wiring up the
-        chain manually."""
-        q = MagicMock()
-        for method in ("select", "is_", "or_", "eq", "limit", "order", "filter"):
-            getattr(q, method).return_value = q
-        q.execute.return_value = MagicMock(data=[])
-        return q
-
-    @pytest.mark.asyncio
-    async def test_keyword_recall_applies_all_lifecycle_filters(self):
-        """_keyword_recall (fallback path) must filter deleted_at, expired_at,
-        and superseded_by server-side — matching the show_history=false
-        branch of match_memories / keyword_search_memories. valid_to is
-        filtered client-side (see test_keyword_recall_filters_past_valid_to)."""
-        from server import _keyword_recall
-
-        query = self._fluent_query()
-        client = MagicMock()
-        client.table.return_value = query
-
-        await _keyword_recall(client, "anything", project="jarvis", mem_type=None, limit=10)
-
-        is_args = [tuple(call.args) for call in query.is_.call_args_list]
-        assert ("deleted_at", "null") in is_args
-        assert ("expired_at", "null") in is_args
-        assert ("superseded_by", "null") in is_args
-
-        # valid_to must NOT appear in .or_() — PostgREST allows only one
-        # `or=` per query, and this function already uses .or_() for project
-        # scoping and for the keyword ILIKE clauses. Putting valid_to in
-        # another .or_() would silently overwrite one of the existing filters.
-        or_filters = [call.args[0] for call in query.or_.call_args_list]
-        assert not any("valid_to" in f for f in or_filters), (
-            f"valid_to must not be pushed into .or_() (PostgREST one-or rule); "
-            f"got or_ calls: {or_filters}"
-        )
-
-        # valid_to must be selected so the client can filter it after fetch.
-        select_args = [call.args[0] for call in query.select.call_args_list]
-        assert any("valid_to" in s for s in select_args), (
-            f"valid_to must be in SELECT for client-side filtering; got select calls: {select_args}"
-        )
-
-    @pytest.mark.asyncio
-    async def test_keyword_recall_brief_mode_also_filters(self):
-        """brief=True takes a different select-columns branch — make sure
-        the filter block still runs and valid_to is still selected."""
-        from server import _keyword_recall
-
-        query = self._fluent_query()
-        client = MagicMock()
-        client.table.return_value = query
-
-        await _keyword_recall(client, "anything", project=None, mem_type=None, limit=5, brief=True)
-
-        is_args = [tuple(call.args) for call in query.is_.call_args_list]
-        assert ("deleted_at", "null") in is_args
-        assert ("expired_at", "null") in is_args
-        assert ("superseded_by", "null") in is_args
-
-        select_args = [call.args[0] for call in query.select.call_args_list]
-        assert any("valid_to" in s for s in select_args)
-
-    @pytest.mark.asyncio
-    async def test_keyword_recall_filters_past_valid_to(self):
-        """Rows whose valid_to is in the past are tombstoned and must be
-        excluded from default recall. Rows with valid_to=None or valid_to
-        in the future are live."""
-        from datetime import datetime, timedelta, timezone
-
-        from server import _keyword_recall
-
-        now = datetime.now(timezone.utc)
-        past = (now - timedelta(days=1)).isoformat()
-        future = (now + timedelta(days=30)).isoformat()
-
-        rows = [
-            {
-                "name": "live_null_vt",
-                "type": "project",
-                "project": "jarvis",
-                "description": "d",
-                "tags": [],
-                "updated_at": now.isoformat(),
-                "valid_to": None,
-            },
-            {
-                "name": "live_future_vt",
-                "type": "project",
-                "project": "jarvis",
-                "description": "d",
-                "tags": [],
-                "updated_at": now.isoformat(),
-                "valid_to": future,
-            },
-            {
-                "name": "tombstoned_past_vt",
-                "type": "project",
-                "project": "jarvis",
-                "description": "d",
-                "tags": [],
-                "updated_at": now.isoformat(),
-                "valid_to": past,
-            },
-        ]
-
-        query = self._fluent_query()
-        query.execute.return_value = MagicMock(data=rows)
-        client = MagicMock()
-        client.table.return_value = query
-
-        result = await _keyword_recall(
-            client, "anything", project="jarvis", mem_type=None, limit=10, brief=True
-        )
-
-        text = result[0].text
-        assert "live_null_vt" in text
-        assert "live_future_vt" in text
-        assert "tombstoned_past_vt" not in text
-        assert "Found 2 memories" in text
-
-    @pytest.mark.asyncio
-    async def test_keyword_recall_all_rows_tombstoned_returns_empty(self):
-        """If every row returned by the DB has a past valid_to, the client-side
-        filter should yield zero live rows and the function should return the
-        empty-result message — not the 'Found N' line with N=0."""
-        from datetime import datetime, timedelta, timezone
-
-        from server import _keyword_recall
-
-        past = (datetime.now(timezone.utc) - timedelta(days=1)).isoformat()
-        rows = [
-            {
-                "name": "dead1",
-                "type": "project",
-                "project": "jarvis",
-                "description": "d",
-                "tags": [],
-                "updated_at": past,
-                "valid_to": past,
-            },
-        ]
-
-        query = self._fluent_query()
-        query.execute.return_value = MagicMock(data=rows)
-        client = MagicMock()
-        client.table.return_value = query
-
-        result = await _keyword_recall(
-            client, "anything", project="jarvis", mem_type=None, limit=10
-        )
-        assert result[0].text == "No memories found."
-
-    @pytest.mark.asyncio
-    async def test_hybrid_recall_defaults_show_history_false(self, monkeypatch):
-        """_hybrid_recall must pass show_history=False to both the semantic
-        and keyword RPCs on the default path — that flag is what the SQL
-        lifecycle filter keys off of (post-#284 fix)."""
-        from server import _hybrid_recall
-
-        client = MagicMock()
-        # Both RPC calls return no data so the function takes the fallback
-        # branch; we only care about what args we sent to the RPCs.
-        client.rpc.return_value.execute.return_value = MagicMock(data=[])
-
-        async def _stub_embed(_text):
-            return [0.0] * 512
-
-        monkeypatch.setattr(server_module, "_embed_query", _stub_embed)
-
-        await _hybrid_recall(
-            client,
-            query_text="anything",
-            project="jarvis",
-            mem_type=None,
-            limit=5,
-        )
-
-        # Both RPCs invoked with show_history=False. Pure-empty result means
-        # recall() bails before any link-expansion RPC, so exactly two calls.
-        rpc_calls = client.rpc.call_args_list
-        assert len(rpc_calls) == 2
-        for call in rpc_calls:
-            rpc_name, rpc_args = call.args[0], call.args[1]
-            assert rpc_args.get("show_history") is False, (
-                f"{rpc_name} called without show_history=False: {rpc_args}"
-            )
-        rpc_names = {call.args[0] for call in rpc_calls}
-        assert "keyword_search_memories" in rpc_names
-        # Semantic RPC is model-dependent (match_memories or match_memories_v2).
-        assert any(name.startswith("match_memories") for name in rpc_names)
-
-    @pytest.mark.asyncio
-    async def test_hybrid_recall_propagates_show_history_true(self, monkeypatch):
-        """When caller opts in to show_history=True, both RPCs must see it —
-        otherwise audit/debug queries can't inspect tombstones."""
-        from server import _hybrid_recall
-
-        client = MagicMock()
-        client.rpc.return_value.execute.return_value = MagicMock(data=[])
-
-        async def _stub_embed(_text):
-            return [0.0] * 512
-
-        monkeypatch.setattr(server_module, "_embed_query", _stub_embed)
-
-        await _hybrid_recall(
-            client,
-            query_text="anything",
-            project=None,
-            mem_type=None,
-            limit=5,
-            show_history=True,
-        )
-
-        for call in client.rpc.call_args_list:
-            assert call.args[1].get("show_history") is True
-
-
-# ---------------------------------------------------------------------------
-# #417: session-snapshot tag filter — recall must drop operational artifacts
-# ---------------------------------------------------------------------------
-
-
-class TestExcludedTagsFilter:
-    """Osasuwu/jarvis#417: session_snapshot_* memories (auto-created by the
-    PreCompact hook) carry mixed transcript content that semantically matches
-    a wide range of queries — they were drowning out real content in recall
-    (q19 top-5 was 5/5 snapshot rows). They're meant to be fetched by name
-    via memory_get during /end recovery, never to compete in normal recall.
-
-    Lock the filter at both layers: the helper itself, and the keyword-recall
-    integration so a future refactor can't silently re-introduce the noise.
-    """
-
-    def test_filter_drops_session_snapshot(self):
-        from server import _filter_excluded_tags
-
-        rows = [
-            {"name": "real_content", "tags": ["pillar-4"]},
-            {"name": "session_snapshot_abc", "tags": ["session-snapshot", "auto"]},
-            {"name": "real_decision", "tags": ["decision", "memory"]},
-        ]
-        out = _filter_excluded_tags(rows)
-        assert [r["name"] for r in out] == ["real_content", "real_decision"]
-
-    def test_filter_preserves_input_order(self):
-        from server import _filter_excluded_tags
-
-        rows = [
-            {"name": "a", "tags": []},
-            {"name": "snap", "tags": ["session-snapshot"]},
-            {"name": "b", "tags": ["x"]},
-            {"name": "c", "tags": None},
-        ]
-        out = _filter_excluded_tags(rows)
-        assert [r["name"] for r in out] == ["a", "b", "c"]
-
-    def test_filter_handles_empty_and_missing_tags(self):
-        from server import _filter_excluded_tags
-
-        # Missing tags key, None tags, empty tags — none should match the
-        # exclude set, all should pass through.
-        rows = [
-            {"name": "no_key"},
-            {"name": "none_tags", "tags": None},
-            {"name": "empty_tags", "tags": []},
-        ]
-        out = _filter_excluded_tags(rows)
-        assert [r["name"] for r in out] == ["no_key", "none_tags", "empty_tags"]
-
-    def test_filter_no_op_on_empty_input(self):
-        from server import _filter_excluded_tags
-
-        assert _filter_excluded_tags([]) == []
-        assert _filter_excluded_tags(None) is None or _filter_excluded_tags(None) == []
-
-    @pytest.mark.asyncio
-    async def test_keyword_recall_excludes_session_snapshots(self):
-        """_keyword_recall must drop session-snapshot rows from the candidate
-        pool before the valid_to client-side filter — otherwise the live
-        budget burns on operational artifacts."""
-        from server import _keyword_recall
-
-        rows = [
-            {
-                "name": "snap1",
-                "type": "project",
-                "project": "jarvis",
-                "description": "d",
-                "content": "c",
-                "tags": ["session-snapshot", "auto"],
-                "updated_at": "2026-04-25T00:00:00+00:00",
-                "valid_to": None,
-            },
-            {
-                "name": "real_mem",
-                "type": "decision",
-                "project": "jarvis",
-                "description": "d",
-                "content": "c",
-                "tags": ["pillar-4"],
-                "updated_at": "2026-04-25T00:00:00+00:00",
-                "valid_to": None,
-            },
-            {
-                "name": "snap2",
-                "type": "project",
-                "project": "jarvis",
-                "description": "d",
-                "content": "c",
-                "tags": ["session-snapshot"],
-                "updated_at": "2026-04-25T00:00:00+00:00",
-                "valid_to": None,
-            },
-        ]
-
-        query = MagicMock()
-        for method in ("select", "is_", "or_", "eq", "limit", "order"):
-            getattr(query, method).return_value = query
-        query.execute.return_value = MagicMock(data=rows)
-        client = MagicMock()
-        client.table.return_value = query
-
-        result = await _keyword_recall(
-            client, "anything", project="jarvis", mem_type=None, limit=10
-        )
-        # Result is a TextContent list — assert real_mem is named, snaps aren't.
-        text = result[0].text if result else ""
-        assert "real_mem" in text
-        assert "snap1" not in text
-        assert "snap2" not in text
-
-
-# ---------------------------------------------------------------------------
-# #286: outcome_record / outcome_update must accept and persist memory_id
-# ---------------------------------------------------------------------------
-
-
-class TestOutcomeMemoryId:
-    """Osasuwu/jarvis#286: task_outcomes.memory_id is the FK that drives the
-    memory_calibration view. The DB column + FK + view already exist; these
-    tests pin the MCP tool layer so future refactors can't drop memory_id
-    from the insert/update payloads and silently re-empty calibration."""
-
-    @pytest.fixture(autouse=True)
-    def _patch_client(self, monkeypatch):
-        self.client = MagicMock()
-        # Chain: client.table(...).insert(...).execute() → data=[{"id": "..."}]
-        self.tbl = MagicMock()
-        self.client.table.return_value = self.tbl
-        self.tbl.insert.return_value.execute.return_value = MagicMock(data=[{"id": "outcome-uuid"}])
-        # For update: client.table(...).update(...).eq(...).execute() → data=[...]
-        self.tbl.update.return_value.eq.return_value.execute.return_value = MagicMock(
-            data=[{"id": "outcome-uuid"}]
-        )
-        monkeypatch.setattr(server_module, "_get_client", lambda: self.client)
-
-    @pytest.mark.asyncio
-    async def test_outcome_record_persists_memory_id(self):
-        from server import _handle_outcome_record
-
-        await _handle_outcome_record(
-            {
-                "task_type": "delegation",
-                "task_description": "test",
-                "outcome_status": "success",
-                "memory_id": "11111111-1111-1111-1111-111111111111",
-            }
-        )
-
-        insert_payload = self.tbl.insert.call_args[0][0]
-        assert insert_payload["memory_id"] == "11111111-1111-1111-1111-111111111111"
-
-    @pytest.mark.asyncio
-    async def test_outcome_record_omits_memory_id_when_absent(self):
-        """Backwards-compat: callers that don't pass memory_id still work,
-        and the key is not written as NULL (keeps existing rows' shape)."""
-        from server import _handle_outcome_record
-
-        await _handle_outcome_record(
-            {
-                "task_type": "delegation",
-                "task_description": "test",
-                "outcome_status": "success",
-            }
-        )
-
-        insert_payload = self.tbl.insert.call_args[0][0]
-        assert "memory_id" not in insert_payload
-
-    @pytest.mark.asyncio
-    async def test_outcome_record_ignores_none_memory_id(self):
-        """Explicit None is treated the same as missing — matches the
-        pattern used for every other optional field in _handle_outcome_record."""
-        from server import _handle_outcome_record
-
-        await _handle_outcome_record(
-            {
-                "task_type": "delegation",
-                "task_description": "test",
-                "outcome_status": "success",
-                "memory_id": None,
-            }
-        )
-
-        insert_payload = self.tbl.insert.call_args[0][0]
-        assert "memory_id" not in insert_payload
-
-    @pytest.mark.asyncio
-    async def test_outcome_update_persists_memory_id(self):
-        from server import _handle_outcome_update
-
-        await _handle_outcome_update(
-            {
-                "id": "outcome-uuid",
-                "memory_id": "22222222-2222-2222-2222-222222222222",
-            }
-        )
-
-        update_payload = self.tbl.update.call_args[0][0]
-        assert update_payload["memory_id"] == "22222222-2222-2222-2222-222222222222"
-
-    @pytest.mark.asyncio
-    async def test_outcome_update_retrolinks_without_status_change(self):
-        """memory_id alone is a valid update — no outcome_status flip required.
-        Covers the backfill flow: /verify already marked the row, we just now
-        discovered which memory informed it."""
-        from server import _handle_outcome_update
-
-        result = await _handle_outcome_update(
-            {
-                "id": "outcome-uuid",
-                "memory_id": "33333333-3333-3333-3333-333333333333",
-            }
-        )
-
-        update_payload = self.tbl.update.call_args[0][0]
-        assert update_payload == {"memory_id": "33333333-3333-3333-3333-333333333333"}
-        # verified_at only auto-sets when outcome_status changes off pending —
-        # memory_id-only update must not touch it.
-        assert "verified_at" not in update_payload
-        assert "updated" in result[0].text
